requirements.txt
fastapi==0.135.3
livekit-agents[google,silero]==1.5.2
livekit-api==1.1.0
python-dotenv==1.2.2
uvicorn==0.44.0
